requirements.txt
langchain>=0.3
langgraph>=0.2
langchain-openai>=0.2
langchain-tavily>=0.1
pydantic>=2.0
python-dotenv>=1.0